Fix ViewPosition

+++ hbs-core/src/main/java/pl/edu/agh/hbs/core/model/ui/cartesian/ViewPosition.java
@@ -3,30 +3,30 @@ package pl.edu.agh.hbs.core.model.ui.cartesian;
 import java.util.Objects;
 
 public class ViewPosition {
-    private double x;
-    private double y;
+    private int x;
+    private int y;
 
     public ViewPosition() {
     }
 
-    public ViewPosition(double x, double y) {
+    public ViewPosition(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
